modify version

+++ ProfitParCountry/src/profitparcountry/ModeleMapReduceDriver.java
@@ -3,6 +3,7 @@ package profitparcountry;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -16,47 +17,7 @@ import org.apache.hadoop.util.ToolRunner;
 
 public class ModeleMapReduceDriver extends Configured implements Tool
 {
-//    @Override
-//    public int run(String[] args) throws Exception
-//    {
-//        // vérifier les paramètres fournis au lancement du mapper
-//        if (args.length != 2) {
-//            System.err.println("Usage: ModeleMapReduceDriver <input path> <outputpath>");
-//            System.exit(-1);
-//        }
-//
-//        // instant de démarrage
-//        Long initTime = System.currentTimeMillis();
-//
-//        // créer le job map-reduce à l'aide de YarnJob qui vérifie sa structure
-//        Configuration conf = this.getConf();
-//        YarnJob job = new YarnJob(conf, "ModeleMapReduce Job");
-//        job.setJarByClass(ModeleMapReduceDriver.class);
-//
-//        // définir les classes Mapper, Reducer et Combiner si nécessaire
-//        job.setMapperClass(ModeleMapReduceMapper.class);
-//        job.setReducerClass(ModeleMapReduceReducer.class);
-//        //job.setCombinerClass(ModeleMapReduceCombiner.class);
-//
-//        // définir les données d'entrée
-//        job.setInputFormatClass(TextInputFormat.class);
-//        job.addInputPath(new Path(args[0]));   /// VOIR ENTREES= DANS LE MAKEFILE
-//        job.setInputDirRecursive(false);       // mettre true si les fichiers sont dans des sous-dossiers
-//
-//        // définir les données de sortie
-//        job.setOutputFormatClass(TextOutputFormat.class);
-//        job.setOutputPath(new Path(args[1])); /// VOIR SORTIE= DANS LE MAKEFILE
-//
-//        // lancer le job et attendre sa fin en comptant les secondes
-//        Long startTime = System.currentTimeMillis();
-//        boolean success = job.waitForCompletion(true);
-//        Long endTime = System.currentTimeMillis();
-//        System.out.println("Job Duration seconds: " + ((endTime-startTime)/1000L));
-//        System.out.println("Total Duration seconds: " + ((endTime-initTime)/1000L));
-//        return success ? 0 : 1;
-//    }
 
-    //au cas où YarnJob ne fonctionne pas, remplacer l'autre méthode par celle-ci
     @Override
     public int run(String[] args) throws Exception
     {
@@ -86,11 +47,11 @@ public class ModeleMapReduceDriver extends Configured implements Tool
 
         // sorties du mapper = entrées du reducer et entrées et sorties du combiner
         job.setMapOutputKeyClass(Text.class);
-        job.setMapOutputValueClass(IntWritable.class);
+        job.setMapOutputValueClass(DoubleWritable.class);
 
     //     sorties du reducer
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(DoubleWritable.class);
 
     //     définir les données de sortie : dossier et types des clés et valeurs
         job.setOutputFormatClass(TextOutputFormat.class);
